refactor: renomeia pacote de interfaces
